+++ source/add_extra_stereo_audio/plugin.py
@@ -38,6 +38,7 @@ class Settings(PluginSettings):
         "channels":              '',
         "codec_name":            '',
         "use_libfdk_aac":         False,
+        "remove_original_multichannel_audio": False,
     }
 
     def __init__(self, *args, **kwargs):
@@ -54,6 +55,9 @@ class Settings(PluginSettings):
             },
             "use_libfdk_aac":               {
                 "label": "check if you want to use libfdk_aac (requires ffmpeg 5), otherwise aac is used",
+            },
+            "remove_original_multichannel_audio":               {
+                "label": "check if you want to remove the original multichannel audio stream",
             }
         }
 
@@ -121,7 +125,10 @@ def on_library_management_file_test(data):
     stream = stream_to_stereo_encode(stream_language, channels, codec_name, probe_streams)
     if stream >= 0:
         data['add_file_to_pending_tasks'] = True
-        logger.debug("Audio stream '{}' is '{}' and has '{}' channels and is encoded with '{}'".format(stream, stream_language, channels, codec_name))
+        if stream_language == '' or channels == '' or codec_name == '':
+            logger.debug("Audio stream '{}' is multichannel audio - convert stream".format(stream))
+        else:
+            logger.debug("Audio stream '{}' is '{}' and has '{}' channels and is encoded with '{}' - convert stream".format(stream, stream_language, channels, codec_name))
     else:
         data['add_file_to_pending_tasks'] = False
         logger.debug("do not add file '{}' to task list - no matching streams".format(abspath))
@@ -180,12 +187,22 @@ def on_worker_process(data):
     codec_name = settings.get_setting('codec_name').lower()
     encoder = 'aac'
     if settings.get_setting('use_libfdk_aac'): encoder = 'libfdk_aac'
+    remove_original = settings.get_setting('remove_original_multichannel_audio')
 
     stream = stream_to_stereo_encode(stream_language, channels, codec_name, probe_streams)
     if stream >= 0:
 
         # Get generated ffmpeg args
-        ffmpeg_args = ['-hide_banner', '-loglevel', 'info', '-i', str(abspath), '-max_muxing_queue_size', '9999', '-map', '0', '-c', 'copy', '-map', '0:a:'+str(stream), '-c:a:'+str(new_audio_stream), encoder, '-ac', '2', '-b:a:'+str(new_audio_stream), '128k', '-y', str(outpath)]
+        if not remove_original:
+            ffmpeg_args = ['-hide_banner', '-loglevel', 'info', '-i', str(abspath), '-max_muxing_queue_size', '9999', '-map', '0', '-c', 'copy', '-map', '0:a:'+str(stream), '-c:a:'+str(new_audio_stream), encoder, '-ac', '2', '-b:a:'+str(new_audio_stream), '128k', '-y', str(outpath)]
+        else:
+            ffmpeg_args = ['-hide_banner', '-loglevel', 'info', '-i', str(abspath), '-max_muxing_queue_size', '9999', '-map', '0:s?', '-c:s', 'copy', '-map', '0:d?', '-c:d', 'copy', '-map', '0:t?', '-c:t', 'copy', '-map', '0:v', '-c:v', 'copy']
+            for astream in range(0, total_audio_streams+1):
+                if astream ==  stream:
+                    ffmpeg_args += ['-map', '0:a:'+str(stream), '-c:a:'+str(stream), encoder, '-ac', '2', '-b:a:'+str(stream), '128k']
+                else:
+                    ffmpeg_args += ['-map', '0:a:'+str(astream), '-c:a:'+str(astream), 'copy']
+            ffmpeg_args += ['-y', str(outpath)]
 
         logger.debug("ffmpeg args: '{}'".format(ffmpeg_args))